requirements.txt
numpy==1.21.6
opencv-python==4.11.0.86
Pillow==9.0.1

# YOLOv8 and dependencies
torch==1.13.1
torchvision==0.14.1
ultralytics==8.0.20

# Web Framework
Flask==2.0.3
Werkzeug==2.0.3
Jinja2==3.0.3
MarkupSafe==2.0.1
itsdangerous==2.0.1

# Machine Learning
scikit-learn==1.0.2

# Utilities
tqdm==4.64.1
matplotlib==3.5.3
pandas==1.3.5
PyYAML==6.0
requests==2.27.1

# Web interface
Flask-SocketIO==5.1.2
python-socketio==5.5.2
python-engineio==4.3.1

# For web scraping
beautifulsoup4==4.10.0 